Refactored DB functions to a separate module so that it is easier to change databases.
