+++ server/db-mongo.js
@@ -1,0 +1,14 @@
+var db = {};
+
+db.getCurrentStatus = function(size, offset) {
+  return [];
+}
+
+db.getHistory = function(size, offset) {
+  return [];
+}
+
+db.upsertStatus = function(deviceId, isOccupied) {
+}
+
+module.exports = db;
